introduce new is_readonly method
This is work towards divorcing the idea of being an lvalue from being writable.
We currently use is_lvalue() to decide whether, e.g., something is valid to
appear on the LHS of an assignment. This is clearly wrong, but we've got by
until now. With some upcoming changes to how function parameters work, this is
going to cause a bit of difficulty interoperating with isundefined. We try to
head off this upcoming mess in this commit and the ones following.