reject function calls that pass read-only values as var parameters
This reverses a previous decision because I realised (a) this is rejected by
CMurphi and (b) this pattern is characteristic of a typo in the model. It is
possible to pass a read-only value as a var parameter and, if we silently accept
this and copy it to temporary storage, the user will probably be surprised by
the outcome. This will require upcoming changes to rumur/src/generate-expr.cc
too.