allow mismatched range types in non-var parameters
It's possible to pass a differently typed range into a non-var parameter:

  var x: 0 .. 10;

  procedure foo(y: 0 .. 5); begin end;
  ...
  foo(x);

However, we were previously rejecting such constructions. There's no good reason
for this and it seems real world models do this, so now we allow it.

We probably need some more validation when generating the handle-copying logic
though.