switch to a lock-free queue
In this commit, we drop the mutex that protects each thread's pending state
queue and implement lock-free algorithms for enqueue and dequeue. The approach
is based on Maged Michael, "Hazard Pointers: Safe Memory Reclamation for
Lock-Free Objects" in TPDS 15(8) 2004. The motivation for this change is to
remove contention on these mutexes late in a checking run when threads are
regularly work-stealing from each other. It is thought that this will improve
scalability and performance on large multiway machines.

Github: related to #52 "mmap queue pages"