fix Clang 5 compatibility
When using Clang 5 on Linux, it picks up GCC's stdatomic.h's definition of
atomic_compare_exchange_strong. This expands to something Clang does not think
is a legal use of an _Atomic type. To avoid this, we revert to the __atomic
built-ins. I think this is LLVM bug https://bugs.llvm.org/show_bug.cgi?id=37457
but it's a little hard to be sure.

This reverts commit a07782b0a17a96c60a74bb978dc5f302055148fd.