add a model of the lock-free queue
This validates some correctness properties about our new queue implementation.
Unfortunately it's not as extensive as it could be because the model size is too
large for my currently available machines. It would be nice to up the number of
threads to 4 and the number of queue nodes to 6, but this seems to exceed the
capacity of what I currently have available.