switch to a non-native package
After some discussion with Nicolas Boulenguez and Sam Hartman it has become
clear to me that it makes more sense to package this as non-native, even without
patches we need to apply on top. Note that to package now, you need to commit a
dummy change that deletes misc/test-debian-packaging.sh because git-buildpackage
does not like diffs from master outside of debian/. Once we've nailed down this
workflow, I'll move that script to master.