add libfl-dev as a build dependency
As identified by Adam Borowski, we depend on this library when we import
FlexLexer.h but we were not declaring it as a build dependency. It worked during
testing because libfl-dev is suggested by Flex, so gets pulled in during a
normal installation of Flex. The Debian build environment however is more strict
and does not pull in suggested packages.