add an interface for asking if the bounds of a type are constant
It's possible to write quantifiers using ad hoc types that don't have constant
bounds. E.g. a range whose upper or lower limit is the value of a local
variable. This is generally not a problem, but for the quantifiers used in
rulesets we require these bounds to be constant in order to know the length of
the loop iteration (the number of contained rules when flattened) in order to
print the rule correctly during counterexample trace. This adds some supporting
code working towards this.

Github: related to #105 "missing parameters in ruleset trace"