fix: report parameter values in counterexample traces through rulesets
This fixes an issue where counterexample traces that traversed rulesets did not
show the values of the parameters to those rulesets. Without these values,
debugging such a counterexample trace is virtually impossible.

I'm not sure whether this was an oversight or intentional misfeature. The
previous code in generate-model.cc (in particular the "<= rule_taken"
comparison) implies that I was aware of the counterexample reporting treating
all ruleset transitions identically, but I'm not sure if I intended this
decision to be a long term behaviour or this was more like an unannotated TODO.

Note that this commit does not fully close #105, because startstates within
rulesets are still reported without their parameter values. Similarly, the
machine readable output shows no parameter values.

Github: related to #105 "missing parameters in ruleset trace"