fix: support passing non-lvalue complex values as parameters
There was a mistaken assumption in the code that generated function/procedure
calls. We were assuming there was no way to construct a complex non-lvalue
expression. It turns out this is actually not true: the return value of a
function can be complex and is a non-lvalue.

Github: closes #109 "assert fail on non-lvalue complex arguments"