implement a value-to-string conversion to avoid PRIVAL
This removes the last assumption of value_t being int64_t. Now (theoretically)
it should be possible to simply typedef value_t to be something other than
int64_t and have everything work. This is part of a longer arc towards allowing
the user (or automatic optimisation) to select a more performant type for
value_t.

Github: related to #23 "Use native scalars of variable size"