allow return of differing range types from functions
It is now possible to, from within a range-returning function, return an
expression of a differing range type. This relaxes some previous limitations in
this area. This is motivated by an upcoming change in which the type of numeric
literals will become non-null.

Github: related to #90 "avoid returning null from Expr::type()"