fix: test for valid negation expressions
This typo seems to have been present more or less since the first implementation
of this code. It seems negative numbers -- or perhaps negation -- are rarely
used in models. The effect of it was that negating a ranged expression would be
rejected and negating a non-ranged expression (including a complex expression)
would be incorrectly accepted.

Github: closes #111 "negation of ranged expressions is rejected"