remove now-unnecessary duplicate overrides
When we were overloading visit(), it was necessary to tell the compiler that
yes, we wanted to retain the ability to override some variants of visit() in
child classes. Now that the visitation methods are all distinct, the compiler
can see this for itself.

Github: related to #89 "stop overloading BaseTraversal::visit"