expose trace() as a macro
f5589751de2f860c3cca7d681f9710160d3c20a8 caused an unexpected performance
regression and debugging this revealed the cause to be related to calls to
trace() being emitted even with tracing disabled. As described in the comment in
this commit, it seems compilers are unwilling to inline calls to varargs
functions or perform interprocedural optimisations through them. This meant that
we would unconditionally incur the overhead of a varargs function call for each
tracing point even with tracing disabled. This was compounded when dynamic
construction of a string was added to some tracing calls.

As a silver lining to this detour, this commit should actually improve
performance beyond the baseline of what we had before the regression, as the
compiler should now be able to elide these calls (and string construction)
altogether.