add a script for packaging Rumur for Debian
Rumur has currently been accepted into Debian's New Queue [0], but there's a
freeze [1] in place for the upcoming Buster release. We got in before the soft
freeze but after the transition freeze, so there's nothing to do but wait.
However, churn on this release script that was in the 'packaging/debian' branch
has more or less stopped, so it seemed a good time to migrate it to 'master'.
Note that to do further Debian packaging, this commit will need to be merged to
'packaging/debian', misc/test-debian-packaging.sh will need to be removed from
that branch, and then we should be able to seamlessly packaging from
'packaging/debian' without juggling files.

Github: related to #85 "release beta"

  [0]: https://ftp-master.debian.org/new/rumur_2019.02.04-1.html
  [1]: https://release.debian.org/buster/freeze_policy.html