implement the remaining cover functionality
Note that cover results are still printed with the index of the cover rather
than its name.

Also, covers within rulesets are treated as a single cover. I.e. it will be
considered hit if any iteration of the ruleset hits it. It's not clear if this
is an intuitive implementation or we should consider changing it in future.

Github: related to #47 "properties, covers, etc"