fix: only decode binary data on-demand in the test suite
It seems when non-ASCII text is present in the data we receive from an external
program, we should have been re-encoding this before writing it back to
sys.stdout or sys.stderr. Instead of doing this, we now don't decode it into
UTF-8 at all to begin with. We later decode stdout only if it is needed (e.g.
for regexing.