use 'bound' to computer state_depth faster when available
When doing bounded model checking (--bound with a non-zero limit), we have a
'bound' member of the state struct. This allows us to directly compute the depth
of a state without having to walk the linked-list of its previous pointers. This
isn't normally a performance-critical path anyway, but there seemed no reason
not to speed this up.