fix: avoid exiting from within exception handlers
Calling exit() from within exception handlers in parse_args causes Valgrind
errors because the destructor for the thrown exception does not get called. By
rearranging this code we can avoid this issue and also be more precise with the
exceptions we catch. That is, we now avoid catching all exceptions from
std::stoul but instead catch only the ones we expect.