ensure handle offsets are only 0-7
We effectively canonicalise handles by ensuring the base always points at the
first byte we would read and the offset is then only the bit offset within this
byte. The purpose of this is to eventually allow reducing the offset member to 3
bits.

Github: related to #44 "drop width member of handle"