fix: use __typeof__ instead of typeof
Some compilers do not treat typeof as a synonym and only recognise their own
built-in, __typeof__ or __typeof. This was not picked up before because it seems
we are not testing on any non-x86-64 platforms. It would be nice to address this
but I don't currently have any other platforms available.