fix: duplicate cex traces on deadlock when using --max-errors > 0
We now use a separate setjmp() for the deadlock check, which resolves a bug
wherein the error() call from the deadlock path would longjmp() back to the main
loop, causing it to simply observe the same deadlock again. Note that this also
coincidentally is a nice step towards removing the `retain` parameter to
error().

Github: closes #121 "deadlock with max errors > 0 causes corruption"
Github: related to #120 "why free states in error()?"