fix: emit an array instead of a string literal for resources
It turns out Clang on Linux finds a corner of the C++ standard to wave at us for
this:

    resources_header.cc:4:3: error:
    string literal of length 93646 exceeds maximum length 65536 that C++
    compilers are required to support [-Werror,-Woverlength-strings]

To avoid this mess, we revert to emitting something closer to what xxd itself
does.

One of the other techniques I explored was avoiding any transformation at all
and just linking the files in as binary blobs (`ld -r -b binary ...`).
Unfortunately it seems it is quite challenging to support this on macOS in a
forwards and backwards compatible way due to changes in the Mach-O format. So an
in-repo script it is.