implement final liveness scan
This isn't the cleanest implementation, but it seems a reasonable first cut.
My initial attempt was at a parallel version of this. I.e. rendezvous all the
threads in exit_with() prior to joining, accumulate error_counts, then check
liveness in parallel if we have no errors. Unfortunately this turned out to be
more complicated than anticipated. The chief problem is that the final liveness
check loop basically keeps running until it is no longer making progress, and
this is not straightforward to implement with multiple threads. In particular, a
repeated non-trivial rendezvous seems awkward with the current structure. By
"non-trivial", I mean having some logic between rendezvous_arrive() and
rendezvous_depart().

Instead of fighting with this to get a maybe-correct implementation, we just go
with a single-threaded approach for now and let's wait until users complain
about the performance. Postponing this work may actually be a win long term
because I am not currently aware of any efficient algorithm for checking such
liveness properties in an explicit state model checker like this. Maybe by the
time this becomes a pain point, we will have such an algorithm. If not, we're
well positioned to go parallel, with all the relevant leaf functions mostly
using atomics where necessary.

Github: related to #47 "properties, covers, etc"