optimise zebra puzzle model
This change effectively pre-assigns people to houses. There was no need to model
this decision within the exploration of the model itself because nothing is
known about the people at the time they are assigned to houses, so this still
leaves the entire problem space open.

This model is still not checkable within 16GB, so it may be an interesting test
case to use for work on compressed representations.