remove the concept of disabled properties
These were included originally because the thinking was that we would eventually
have command line options allowing the user to change the types of individual
properties. In this scheme, one of the possibilities would be "disable this
property." Now that all the planned property types are implemented (invariants,
covers, and liveness) it's become clear that changing the type of a property
after it's been written isn't something you generally want to do. Moreover, if
you really do want to do this it's unlikely something you need to be tweakable
from the command line; you can just go into the model source and edit the
relevant property.

Note that a side effect of this change is that 'property' is no longer a
keyword.

Github: closes #47 "properties, covers, etc"