fix: correct previous value read when determining deltas to print
When printing a diff counter-example trace (the default), some values that did
not change were being re-printed. That is, sometimes there would be no delta in
the value of a state component between one state and the next but its value
would be printed anyway. The underlying problem was that the handle used to read
from the previous state was incorrect. More specifically, this handle was always
using a base value of the start of the state data rather than the byte in which
the current field started.

This bug was introduced in d783655eae837b805b69185d1d198ea142825973 when state
handles were changed to use a closer base. We missed that this left the
calculation of the handle to the previous state in state_print incorrect.

Github: closes #125 "diff traces over share with arrays"