optimisation: don't require a jmp_buf when --max-errors < 2
We were using the sigsetjmp()/siglongjmp() exception mechanism when either of
two things were true:

  1. --max-errors > 1; or
  2. the model has assumptions.

As of 7dda120345da13f739427915fde630d71bae9ff5, a failed assumption is signalled
via a normal return value, not via a siglongjmp(). As a result, we no longer
need to require a jmp_buf in this case. This should result in a slight speed up
for models that have assumptions but run with --max-errors < 2.

Github: related to #127 "--max-errors > 1 produces unsafe code"