fix syntax error misaligning caret indicator
When your input model contains a syntax error and the line in which the error
occurs contains a tab, Rumur would print an error message indicating the wrong
column location in the line. We now account for tabs and offset the caret
indicator correctly.

Note that error messages for lines containing multibyte characters (e.g. "≔")
are still misaligned, but it does not seem worth it to work around this as it is
fixed upstream in Bison 3.4.

Github: closes #129 "syntax error output is misaligned with tabs"