fix: decode data received from subprocess as UTF-8 in the test suite
This sort of all just worked when using a UTF-8 locale, but when your locale is
not setup this way exceptions occur. The problem is that we should have been
converting the binary data we received from a subprocess to a string immediately
if we intended to then go on and treat it like a string.

Github: related to #134 "Tests fail: UnicodeDecodeError: 'ascii' codec can't
  decode byte 0xe2 in position 120: ordinal not in range(128)"