fix compilation on platforms where size_t != unsigned long
The make_ul function and its usage was mistakenly assuming that size_t was a
typedef for unsigned long. This is not true on, e.g., i386 FreeBSD. This caused
a compilation failure when building Rumur. Instead of trying to fix this
locally, we turn most of these option variables into GMP mpz_classes. Given that
the platform the verifier is generated on may not be the same as the platform it
is compiled on, it is easier to think of most of these values as unbounded
numbers.

Github: related to #136 "Build fails on i386 systems: error: no matching
  function for call to 'make_ul'"