remove the 'action' parameter from rendezvous functions
We were always passing the set_update function for this parameter, so just hard
code a call to this. The larger motivation for this is that we want to move the
rendezvous opt out to an earlier location in header.c. This backs out part of
8cf9d785c925554e6ec4b2a8a55e619f3ecc66f2.

Github: related to #131 "minimal trace mode"