move rendezvous opt out out of exit_with
The purpose of this seemingly irrelevant change is to ensure we opt out of the
rendezvous protocol as soon as error() is called when MAX_ERRORS == 1. In
upcoming changes to ensure we get the shortest resulting trace, we're going to
loop indefinitely in error(). This is only possible without livelocking if we're
no longer part of the thread rendezvous protocol.

Github: related to #131 "minimal trace mode"