suppress output of -m32 check in test suite
When using a toolchain without 32-bit support, the check that eventually skips
this test was causing some output about missing headers that could look
confusing when running the test suite. Even on macOS where the test is not
skipped, it was producing a linker warning. We now pipe all output to /dev/null
as it is not informative for the user.