descend into referents during validation
We're about to introduce a function for validating any subpart of the AST. That
is, effectively calling validate_model with a node other than the topmost node.
This means that the referent of a reference node (ExprID, FunctionCall or
TypeExprID) may not yet have been validated. So we need to drop the optimization
that avoids descending into these.

This makes validating a model take longer, but the difference should be
negligible. If this becomes a problem, we could start memoising visitor
dispatch.