restructure smt::Logic
In the move to support arrays, it looked like we would have to introduce some
multiple inheritance to avoid duplicating code. Instead of doing this, the
present commit removes inheritance entirely and uses a single Logic class. A
pleasant side effect of this change is that neither integer nor bitvector
support is required if your model only uses booleans. E.g. you can use QF_IDL
because integer or bitvector support is only required on demand when an
expression needing them is encountered.