support array TypeDecls in the SMT bridge
Since we inline TypeExprs at the site of their usage, all that is needed for
array support in VarDecls was to remove the Unsupported stubs. Note that the
same has been done for records but we have no ability to translate Records into
SMT yet. We still need support for arrays in expressions to complete array
support in the SMT bridge.