define types in SMT problem
We now expose types to the SMT bridge when constructing a problem for it. That
is, a TypeDecl in Rumur becomes an SMT sort. These are not yet used for, e.g.,
VarDecl generation but they will be soon. The overarching motivation for this is
support for record types, that we can't easily do without this.

Github: related to #150 "define types to SMT bridge"