allow a return type to reference a type that is shadowed by a function parameter
This change permits something like the following to now work:

  type t: 0 .. 1;

  function foo(t: boolean): t; begin
  ...

Previously, the parameter to the function would shadow the named type,
preventing the returned type from being resolved. This was unnecessary, as
there's no way for a return type to depend on a function parameter, so it
doesn't make sense to have the parameter in scope when resolving the return
type.